+++ src/components/global/AppBar/index.ts
@@ -1,0 +1,1 @@
+export { default as AppBar } from "./AppBar";
