+++ src/components/global/CustomButton/index.ts
@@ -1,0 +1,1 @@
+export { default as CustomButton } from "./CustomButton";
