+++ src/components/global/CustomInput/index.ts
@@ -1,0 +1,1 @@
+export { default as CustomInput } from "./CustomInput";
