+++ src/components/global/ProgressBar/ProgressBar.tsx
@@ -1,6 +1,5 @@
 // ProgressBar.tsx
-import React from 'react';
-import styled from 'styled-components';
+import { Container, ProgressLabel, ProgressBarContainer, Progress } from './ProgressBar.styles';
 
 interface ProgressBarProps {
   /**
@@ -33,34 +32,7 @@ interface ProgressBarProps {
   fontSize?: string;
 }
 
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-`;
-
-const ProgressLabel = styled.span<{ textColor: string; fontSize: string }>`
-  color: ${props => props.textColor};
-  font-size: ${props => props.fontSize};
-  font-weight: 500;
-`;
-
-const ProgressBarContainer = styled.div<{ width: string; height: string; backgroundColor: string }>`
-  width: ${props => props.width};
-  height: ${props => props.height};
-  background-color: ${props => props.backgroundColor};
-  border-radius: 100px;
-  overflow: hidden;
-`;
-
-const Progress = styled.div<{ progress: number; progressColor: string }>`
-  width: ${props => `${props.progress}%`};
-  height: 100%;
-  background-color: ${props => props.progressColor};
-  transition: width 0.3s ease-in-out;
-`;
-
-export function ProgressBar({
+export default function ProgressBar({
   progress,
   width = '300px',
   height = '8px',
@@ -73,10 +45,7 @@ export function ProgressBar({
 
   return (
     <Container>
-      <ProgressLabel 
-        textColor={textColor}
-        fontSize={fontSize}
-      >
+      <ProgressLabel textColor={textColor} fontSize={fontSize}>
         {normalizedProgress}% 진행
       </ProgressLabel>
       <ProgressBarContainer
@@ -84,10 +53,7 @@ export function ProgressBar({
         height={height}
         backgroundColor={backgroundColor}
       >
-        <Progress
-          progress={normalizedProgress}
-          progressColor={progressColor}
-        />
+        <Progress progress={normalizedProgress} progressColor={progressColor} />
       </ProgressBarContainer>
     </Container>
   );
