+++ src/components/global/ProgressBar/index.ts
@@ -1,0 +1,1 @@
+export { default as ProgressBar } from "./ProgressBar";
