+++ src/components/global/ProgressBlock/index.ts
@@ -1,0 +1,1 @@
+export { default as ProgressBlock } from "./ProgressBlock";
