+++ src/components/global/SocialLoginButton/index.ts
@@ -1,0 +1,1 @@
+export { default as SocialLoginButton } from "./SocialLoginButton";
