+++ src/components/global/StepProgress/index.ts
@@ -1,0 +1,1 @@
+export { default as StepProgress } from "./StepProgress";
