+++ src/components/global/SubMenu/SubMenu.tsx
@@ -1,30 +1,23 @@
-import { ReactNode } from 'react';
-import { Arrow } from '../../../assets/svg'; 
-import { useNavigate } from 'react-router-dom'; 
-import styled from 'styled-components';
-import theme from '../../../style/theme'; 
- export interface SubMenuProps {
-  text: string; // 공지사항 텍스트
-  to: string; // 이동할 페이지 URL 
+// SubMenu.tsx
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { Arrow } from "../../../assets/svg";
+import { StyledSubMenu, Text } from "./SubMenu.styles";
+
+export interface SubMenuProps {
+  /**
+   * 공지사항 텍스트
+   */
+  text: string;
+  /**
+   * 이동할 페이지 URL
+   */
+  to: string;
 }
 
-const StyledSubMenu = styled.div`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  font-size: 16px;
-  color: ${theme.colors.black100};
-  justify-content: space-between;  
-`;
+export default function SubMenu({ text, to }: SubMenuProps) {
+  const navigate = useNavigate();
 
-const Text = styled.span`
-  white-space: nowrap; 
-  font-size: 16px;
-  color: ${theme.colors.black100};
-`;
-
-export function SubMenu({ text, to }: SubMenuProps) {
-  const navigate = useNavigate(); 
   const handleClick = () => {
     navigate(to); // to에 지정된 URL로 이동
   };
