+++ src/components/global/SubMenu/index.ts
@@ -1,0 +1,1 @@
+export { default as SubMenu } from "./SubMenu";
