+++ src/components/global/Tag/Tag.tsx
@@ -1,9 +1,11 @@
-import styled from "styled-components";
-import { colors } from "../../../style/color";
-import { typography } from "../../../style/typography";
+// Tag.tsx
+import React from "react";
+import { TagWrapper } from "./Tag.styles";
 
 interface TagProps {
-  // tagName은 이제 children으로 대체됩니다.
+  /**
+   * 스타일 옵션: 색상 지정
+   */
   color?: {
     backgroundColor?: string;
     borderColor?: string;
@@ -11,16 +13,6 @@ interface TagProps {
   };
 }
 
-const TagWrapper = styled.div<TagProps>`
-  color: ${({ color }) => color?.fontColor || colors.blue200}; /* 폰트 색상 */
-  border: 1px solid ${({ color }) => color?.borderColor || colors.blue200}; /* border 색상 */
-  border-radius: 5px;
-  background-color: ${({ color }) => color?.backgroundColor || colors.blue100}; /* 배경 색상 */
-  font-size: ${typography.body6};
-  padding: 3px 10px;
-  font-weight: 500;
-`;
-
-export function Tag({ children, color }: React.PropsWithChildren<TagProps>) {
+export default function Tag({ children, color }: React.PropsWithChildren<TagProps>) {
   return <TagWrapper color={color}>{children}</TagWrapper>;
 }
