+++ src/components/global/Tag/index.ts
@@ -1,0 +1,1 @@
+export { default as Tag } from "./Tag";
